Create FileHandler file
